backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
google-api-python-client==2.153.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0


